Update usage of getvar for puppet 6 compatibility
With puppet6, getvar was made a core function and replaces stdlib
version. The behaviour is slightly different with respect to dots within
the key name, which are now treated as data structure traversal
operators and thus the `letsencrypt_csr_${domain}` fact lookup no longer
works.

This commit changes the lookup in a backward compatible way to find the
csr fact under puppet6.

Fixes #42